server/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
asyncpg>=0.24.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0